Refactoring out useful redirect functionality for web driver to use and a failing test for web driver following redirects
